+++ module_1/exercises_rest_express/the-good-corner/backend/src/schemas/ad-schema.ts
@@ -3,10 +3,11 @@ import Joi from "joi";
 const adSchema = Joi.object({
   title: Joi.string().required(),
   description: Joi.string().optional(),
-  owner: Joi.string().optional(),
-  price: Joi.number().optional(),
+  owner: Joi.string().required(),
+  price: Joi.number().required(),
   picture: Joi.string().optional(),
   location: Joi.string().optional(),
+  createdAt: Joi.string().required(),
 });
 
 export default adSchema;