+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- Introduction backend build database and connect app in database ==> { commit: 376ffde8d9046dce8d3d5ba0015bbcee4f24bea8 }
 - Introduction backend build api rest with express ==> { commit: 6b8870829a03cb4127e5e1df6aacfd02af5751e1 }
 - Introduction in TypeScript ==> { commit: cd5cb61eed01457a13cc35b7dfe43372373a032f }
 - Initial Commit ==> { commit: 62b772456377a9311de29ecf2135f89a7fec45a5 }